# refactor: optimizing dropoutEnrollment command