… public/duo-streak.png
… public/skill-card.png